create db model
