+++ src/test/java/de/bwvaachen/botscheduler/TestDBModell.java
@@ -1,0 +1,4 @@
+package de.bwvaachen.botscheduler;
+
+public class TestDBModell {
+}
